type fix

+++ main_mfnet_gtea.py
@@ -68,7 +68,7 @@ def train_cnn_mo(model, optimizer, criterion, train_iterator, num_outputs, use_g
         targets = targets.cuda(gpus[0]).transpose(0,1) # needs transpose to get the first dim to be the task and the second dim to be the batch
 
         if use_gaze or use_hands:
-            targets = targets[num_outputs, :].astype(np.int64)
+            targets = targets[num_outputs, :].astype(np.int64).long()
         assert len(targets) == num_outputs
 
         losses_per_task = []
@@ -145,7 +145,7 @@ def test_cnn_mo(model, criterion, test_iterator, num_outputs, use_gaze, use_hand
             targets = targets.cuda(gpus[0]).transpose(0, 1)
 
             if use_gaze or use_hands:
-                targets = targets[num_outputs, :].astype(np.int64)
+                targets = targets[num_outputs, :].long()
             assert len(targets) == num_outputs
 
             losses_per_task = []
